requirements.txt
asgiref==3.2.3
asn1crypto==0.24.0
Babel==2.8.0
bcrypt==3.1.7
cffi==1.12.2
cryptography==2.6.1
dj-database-url==0.4.1
Django==2.1.5
django-admin-honeypot==1.1.0
django-appconf==1.0.4
django-argonauts==1.2.0
django-axes==5.3.1
django-crispy-forms==1.8.1
django-debug-toolbar==1.11
django-formtools==2.2
django-ipware==2.1.0
django-mfa==2.0
django-otp==0.8.1
django-phonenumber-field==3.0.1
django-ranged-response==0.2.0
django-session-timeout==0.1.0
django-simple-captcha==0.5.12
django-two-factor-auth==1.11.0
gunicorn==19.6.0
phonenumberslite==8.12.0
Pillow==7.0.0
psycopg2==2.7
psycopg2-binary==2.8.5
pycparser==2.19
python-u2flib-server==5.0.0
pytz==2018.9
qrcode==6.1
six==1.12.0
sqlparse==0.3.0
whitenoise==3.1
